requirements.txt

django-next-url-mixin
django-simple-captcha
# Additional requirements go here
